refactor: remove `internal_` prefix from unexported func name
See: BEDS-1065
